Go back to Koffi 2 in master branch
Koffi 3 is languishing and there are too many incompatibilities,
so instead I will progressively pick some stuff from it and improve
Koffi 2 progressively.

+++ src/koffi/test/union.c
@@ -106,5 +106,5 @@ EXPORT void MakeMultiUIndirect(float a, float b, MultiU *out)
     out->f2[1] = b;
 }
 
-EXPORT float GetMultiDouble(MultiU u) { return (float)u.d; }
-EXPORT float GetMultiUnsigned(MultiU u) { return (float)u.u; }
+EXPORT float GetMultiDouble(MultiU u) { return u.d; }
+EXPORT float GetMultiUnsigned(MultiU u) { return u.u; }
